Notes and exercises for the grids & images section (#27)
All material for the grids and images part is in the grids folder. The README
file has the notes that I'll use in place of slides. Students will be pointed
to this since there are links in there for further reading on the GMT docs.
I'll quickly introduce a section with the README on the screen and then jump to
VSCode to do some exercises (starting with the defaults and then adding some
customization on top of it). The final part is a group exercise for them to
make a relief map of a country of their choice.